fix rating
